+++ src/enumerator.rs
@@ -88,9 +88,8 @@ unsafe impl private::ReprValue for Enumerator {
 impl ReprValue for Enumerator {}
 
 impl TryConvert for Enumerator {
-    #[inline]
-    fn try_convert(val: &Value) -> Result<Self, Error> {
-        Self::from_value(*val).ok_or_else(|| {
+    fn try_convert(val: Value) -> Result<Self, Error> {
+        Self::from_value(val).ok_or_else(|| {
             Error::new(
                 exception::type_error(),
                 format!("no implicit conversion of {} into Enumerator", unsafe {