+++ src/integer.rs
@@ -381,9 +381,8 @@ unsafe impl private::ReprValue for Integer {
 impl ReprValue for Integer {}
 
 impl TryConvert for Integer {
-    #[inline]
-    fn try_convert(val: &Value) -> Result<Self, Error> {
-        match Self::from_value(*val) {
+    fn try_convert(val: Value) -> Result<Self, Error> {
+        match Self::from_value(val) {
             Some(i) => Ok(i),
             None => protect(|| {
                 debug_assert_value!(val);