+++ src/module.rs
@@ -169,9 +169,8 @@ unsafe impl private::ReprValue for RModule {
 impl ReprValue for RModule {}
 
 impl TryConvert for RModule {
-    #[inline]
-    fn try_convert(val: &Value) -> Result<Self, Error> {
-        Self::from_value(*val).ok_or_else(|| {
+    fn try_convert(val: Value) -> Result<Self, Error> {
+        Self::from_value(val).ok_or_else(|| {
             Error::new(
                 exception::type_error(),
                 format!("no implicit conversion of {} into Module", unsafe {