+++ src/r_complex.rs
@@ -68,9 +68,8 @@ unsafe impl private::ReprValue for RComplex {
 impl ReprValue for RComplex {}
 
 impl TryConvert for RComplex {
-    #[inline]
-    fn try_convert(val: &Value) -> Result<Self, Error> {
-        Self::from_value(*val).ok_or_else(|| {
+    fn try_convert(val: Value) -> Result<Self, Error> {
+        Self::from_value(val).ok_or_else(|| {
             Error::new(
                 exception::type_error(),
                 format!("no implicit conversion of {} into Complex", unsafe {