+++ src/r_file.rs
@@ -70,9 +70,8 @@ unsafe impl private::ReprValue for RFile {
 impl ReprValue for RFile {}
 
 impl TryConvert for RFile {
-    #[inline]
-    fn try_convert(val: &Value) -> Result<Self, Error> {
-        Self::from_value(*val).ok_or_else(|| {
+    fn try_convert(val: Value) -> Result<Self, Error> {
+        Self::from_value(val).ok_or_else(|| {
             Error::new(
                 exception::type_error(),
                 format!("no implicit conversion of {} into File", unsafe {