+++ src/range.rs
@@ -335,9 +335,8 @@ unsafe impl private::ReprValue for Range {
 impl ReprValue for Range {}
 
 impl TryConvert for Range {
-    #[inline]
-    fn try_convert(val: &Value) -> Result<Self, Error> {
-        Self::from_value(*val).ok_or_else(|| {
+    fn try_convert(val: Value) -> Result<Self, Error> {
+        Self::from_value(val).ok_or_else(|| {
             Error::new(
                 exception::type_error(),
                 format!("no implicit conversion of {} into Range", unsafe {