+++ src/symbol.rs
@@ -231,9 +231,8 @@ unsafe impl private::ReprValue for Symbol {
 impl ReprValue for Symbol {}
 
 impl TryConvert for Symbol {
-    #[inline]
-    fn try_convert(val: &Value) -> Result<Self, Error> {
-        Self::from_value(*val).ok_or_else(|| {
+    fn try_convert(val: Value) -> Result<Self, Error> {
+        Self::from_value(val).ok_or_else(|| {
             Error::new(
                 exception::type_error(),
                 format!("no implicit conversion of {} into Symbol", unsafe {